# ORlab2/podaci/ORLAB3/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "info": "Vrt",
    "version": "1.0"
  },
  "paths": {
    "/product/{sifra}": {
      "get": {
        "summary": "Get resource by ID",
        "description": "Returns a single collection resource",
        "parameteres": [
          {
            "name": "sifra",
            "in": "path",
            "description": "ID of a resource",
            "required": "True",
            "type": "Integer",
            "format": "int64"
          },
          {
            "name": "naziv",
            "in": "collection",
            "description": "Name of the plant",
            "required": "True",
            "type": "String",
            "format": "byte"
          },
          {
            "name": "wiki",
            "in": "collection",
            "description": "Wiki handle for plant's wiki page",
            "required": "True",
            "type": "String",
            "format": "byte"
          },
          {
            "name": "latnaziv",
            "in": "collection",
            "description": "Latin name for a plant",
            "required": "True",
            "type": "String",
            "format": "byte"
          }
        ],
        "responses": {
          "200": {
            "description": "Fetched vrt entry"
          },
          "400": {
            "description": "Plant with provided ID does not exist"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a database entry",
        "description": "Deletes an entry with provided ID",
        "parameteres": [
          {
            "name": "sifra",
            "in": "path",
            "description": "ID of a resource",
            "required": "True",
            "type": "Integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Entry was deleted"
          },
          "400": {
            "description": "entry with provided ID does not exist"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      },
      "put": {
        "summary": "Update resource by ID",
        "description": "Updates resource by ID",
        "parameteres": [
          {
            "name": "sifra",
            "in": "path",
            "description": "ID of a resource",
            "required": "True",
            "type": "Integer",
            "format": "int64"
          },
          {
            "name": "naziv",
            "in": "collection",
            "description": "Name of the plant",
            "required": "True",
            "type": "String",
            "format": "byte"
          },
          {
            "name": "wiki",
            "in": "collection",
            "description": "Wiki handle for plant's wiki page",
            "required": "True",
            "type": "String",
            "format": "byte"
          },
          {
            "name": "latnaziv",
            "in": "collection",
            "description": "Latin name for a plant",
            "required": "True",
            "type": "String",
            "format": "byte"
          }
        ],
        "responses": {
          "200": {
            "description": "Updated vrt entry"
          },
          "400": {
            "description": "Entry with provided ID does not exist"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      }
    },
    "/product": {
      "post": {
        "summary": "Post entry into database",
        "description": "Creates desired entry in database",
        "parameteres": [
          {
            "name": "sifra",
            "in": "path",
            "description": "ID of a resource",
            "required": "True",
            "type": "Integer",
            "format": "int64"
          },
          {
            "name": "naziv",
            "in": "collection",
            "description": "Name of the plant",
            "required": "True",
            "type": "String",
            "format": "byte"
          },
          {
            "name": "wiki",
            "in": "collection",
            "description": "Wiki handle for plant's wiki page",
            "required": "True",
            "type": "String",
            "format": "byte"
          },
          {
            "name": "latnaziv",
            "in": "collection",
            "description": "Latin name for a plant",
            "required": "True",
            "type": "String",
            "format": "byte"
          }
        ],
        "responses": {
          "200": {
            "description": "Added vrt entry"
          },
          "400": {
            "description": "Please provide all parameteres"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      },
      "get": {
        "summary": "Get all entries",
        "description": "Fetches all entries in database",
        "responses": {
          "200": {
            "description": "Got all entries"
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": "Resource not found"
          }
        }
      }
    }
  }
}